# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/me": {
      "get": {
        "description": "Проверяет валидность токена",
        "operationId": "Me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Пользователь авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Проверка авторизации",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/signIn": {
      "post": {
        "operationId": "signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "strongPassword123"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Вход по логину и паролю",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/signUp": {
      "post": {
        "description": "Создаёт нового пользователя и возвращает accessToken в cookie",
        "operationId": "signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "strongPassword123"
                  },
                  "name": {
                    "type": "string",
                    "example": "Алексей"
                  }
                },
                "required": [
                  "email",
                  "password",
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Пользователь успешно зарегистрирован",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "user": {
                      "id": 1,
                      "email": "user@example.com",
                      "firstName": "John",
                      "lastName": "Doe"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Невалидные данные (email уже существует или пароль слишком короткий)"
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        },
        "summary": "Регистрация нового пользователя",
        "tags": [
          "Auth"
        ]
      }
    },
    "/products": {
      "post": {
        "operationId": "ProductsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    },
    "/products/getAll": {
      "get": {
        "description": "Получение всех товаров",
        "operationId": "getAllProducts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Все товары получены",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                },
                "example": [
                  {
                    "id": 1,
                    "name": "Ноутбук Pro",
                    "description": "Мощный ноутбук для работы и развлечений",
                    "price": "29990",
                    "image": "https://img.freepik.com/premium-photo/"
                  }
                ]
              }
            }
          },
          "400": {
            "description": "Ошибка получения "
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        },
        "summary": "Получение товаров",
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "description": "Получение товара",
        "operationId": "getProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Товар получен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                },
                "example": {
                  "id": 1,
                  "name": "Ноутбук Pro",
                  "description": "Мощный ноутбук для работы и развлечений",
                  "price": "29990",
                  "image": "https://img.freepik.com/premium-photo/"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка получения"
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        },
        "summary": "Получение товара",
        "tags": [
          "Products"
        ]
      },
      "patch": {
        "operationId": "ProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      },
      "delete": {
        "operationId": "ProductsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Products"
        ]
      }
    }
  },
  "info": {
    "title": "Todo example",
    "description": "The Todo API description",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "Todo",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "ID пользователя"
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Email пользователя"
          },
          "firstName": {
            "type": "string",
            "example": "Alex",
            "description": "Имя пользователя"
          },
          "picture": {
            "type": "string",
            "example": "https://picture.com",
            "description": "Имя пользователя"
          },
          "lastName": {
            "type": "string",
            "example": "Budai",
            "description": "Фамилия пользователя"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "Budai",
            "description": "Фамилия пользователя"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "Budai",
            "description": "Фамилия пользователя"
          }
        },
        "required": [
          "id",
          "email"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {}
      },
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "ID товара"
          },
          "name": {
            "type": "string",
            "example": "Смартфон XYZ",
            "description": "Имя товара"
          },
          "description": {
            "type": "string",
            "example": "Современный смартфон с отличной камерой и быстрым процессором",
            "description": "Описание товара"
          },
          "price": {
            "type": "string",
            "example": "29990",
            "description": "Цена"
          },
          "image": {
            "type": "string",
            "example": "https://img.freepik.com/p",
            "description": "Фото товара"
          }
        },
        "required": [
          "id",
          "name",
          "price"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}